TEIID-3652 promoting the outer join logic to a separate rule # Conflicts: #	engine/src/main/java/org/teiid/query/optimizer/relational/rules/RuleConstants.java
