+++ engine/src/main/java/org/teiid/query/optimizer/relational/rules/RulePlanJoins.java
@@ -171,162 +171,11 @@ public class RulePlanJoins implements OptimizerRule {
                 continue;
             }
             
-            if (bestOrder.length > 3) {
-        		//bushify
-        		//throw new AssertionError();
-        	}
-                        
             joinRegion.changeJoinOrder(bestOrder);
             joinRegion.reconstructJoinRegoin();
         }
-        
-        planLeftOuterJoinAssociativity(plan, metadata, capabilitiesFinder,
-				analysisRecord, context);
-                
+                        
         return plan;
-    }
-
-    /*
-     * TODO: this is not yet done deeply.  It runs at a single level and then rule raise access runs.
-	 * if there are multiple are several stages of such rearrangements, then we should consider breaking 
-	 * this off as it's own rule after rule raise access that performs raise and keeps running until the
-	 * plan is stable
-     */
-	private void planLeftOuterJoinAssociativity(PlanNode plan,
-			QueryMetadataInterface metadata,
-			CapabilitiesFinder capabilitiesFinder,
-			AnalysisRecord analysisRecord, CommandContext context)
-			throws QueryMetadataException, TeiidComponentException {
-		
-    	LinkedHashSet<PlanNode> joins = new LinkedHashSet<PlanNode>(NodeEditor.findAllNodes(plan, NodeConstants.Types.JOIN, NodeConstants.Types.ACCESS)); 
-    	while (!joins.isEmpty()) {
-    		Iterator<PlanNode> i = joins.iterator();
-    		PlanNode join = i.next();
-    		i.remove();
-    		if (!join.getProperty(Info.JOIN_TYPE).equals(JoinType.JOIN_LEFT_OUTER)) {
-    			continue;
-    		}
-    		PlanNode childJoin = null;
-    		PlanNode other = null;
-    		PlanNode left = join.getFirstChild();
-    		PlanNode right = join.getLastChild();
-
-    		if (isLeftOuterJoin(left)) {
-    			childJoin = left;
-    			other = right;
-    		} else if (isLeftOuterJoin(right)) {
-    			childJoin = right;
-    			other = left;
-    		} else {
-    			continue;
-    		}
-    		
-    		PlanNode cSource = FrameUtil.findJoinSourceNode(other);
-    		
-			if (cSource.getType() != NodeConstants.Types.ACCESS) {
-				continue;
-			}
-    		
-    		List<Criteria> joinCriteria = (List<Criteria>) join.getProperty(Info.JOIN_CRITERIA);
-    		if (!isCriteriaValid(joinCriteria, metadata, join)) {
-    			continue;
-    		}
-    		
-    		List<Criteria> childJoinCriteria = (List<Criteria>) childJoin.getProperty(Info.JOIN_CRITERIA);
-    		if (!isCriteriaValid(childJoinCriteria, metadata, childJoin)) {
-    			continue;
-    		}
-    		
-    		//there are three forms we can take
-    		// (a b) c -> a (b c) or (a c) b
-    		// c (b a) -> (c b) a or (c a) b
-    		Set<GroupSymbol> groups = GroupsUsedByElementsVisitor.getGroups(joinCriteria);
-			if (Collections.disjoint(groups, FrameUtil.findJoinSourceNode(childJoin == left?childJoin.getFirstChild():childJoin.getLastChild()).getGroups())) {
-				//case where absolute order remains the same
-				PlanNode bSource = FrameUtil.findJoinSourceNode(childJoin == left?childJoin.getLastChild():childJoin.getFirstChild());
-				if (bSource.getType() != NodeConstants.Types.ACCESS) {
-    				continue;
-    			}
-    			if (RuleRaiseAccess.canRaiseOverJoin(childJoin == left?Arrays.asList(bSource, cSource):Arrays.asList(cSource, bSource), metadata, capabilitiesFinder, joinCriteria, JoinType.JOIN_LEFT_OUTER, analysisRecord, context, false) == null) {
-    				continue;
-    			}
-    			//rearrange
-    			PlanNode newParent = createJoinNode();
-    			newParent.setProperty(Info.JOIN_TYPE, JoinType.JOIN_LEFT_OUTER);
-    			PlanNode newChild = createJoinNode();
-    			newChild.setProperty(Info.JOIN_TYPE, JoinType.JOIN_LEFT_OUTER);
-    			joins.remove(childJoin);
-    			if (childJoin == left) {
-    				//a (b c)
-    				newChild.addFirstChild(childJoin.getLastChild());
-    				newChild.addLastChild(other);
-    				newChild.setProperty(Info.JOIN_CRITERIA, joinCriteria);
-    				newParent.addFirstChild(childJoin.getFirstChild());
-    				newParent.addLastChild(newChild);
-    				newParent.setProperty(Info.JOIN_CRITERIA, childJoinCriteria);
-    			} else {
-    				//(c b) a
-    				newChild.addFirstChild(other);
-    				newChild.addLastChild(childJoin.getFirstChild());
-    				newChild.setProperty(Info.JOIN_CRITERIA, joinCriteria);
-    				newParent.addFirstChild(newChild);
-    				newParent.addLastChild(childJoin.getLastChild());
-    				newParent.setProperty(Info.JOIN_CRITERIA, childJoinCriteria);
-    			}
-				join.getParent().replaceChild(join, newParent);
-    		} else if (Collections.disjoint(groups, FrameUtil.findJoinSourceNode(childJoin == right?childJoin.getFirstChild():childJoin.getLastChild()).getGroups())) {
-				PlanNode aSource = FrameUtil.findJoinSourceNode(childJoin == left?childJoin.getFirstChild():childJoin.getLastChild());
-				if (aSource.getType() != NodeConstants.Types.ACCESS) {
-    				continue;
-    			}
-
-    			if (!join.getExportedCorrelatedReferences().isEmpty()) {
-        			//TODO: we are not really checking that specifically
-    				continue;
-    			}
-    			if (RuleRaiseAccess.canRaiseOverJoin(childJoin == left?Arrays.asList(aSource, cSource):Arrays.asList(cSource, aSource), metadata, capabilitiesFinder, joinCriteria, JoinType.JOIN_LEFT_OUTER, analysisRecord, context, false) == null) {
-    				continue;
-    			}
-    			
-    			//rearrange
-    			PlanNode newParent = createJoinNode();
-    			newParent.setProperty(Info.JOIN_TYPE, JoinType.JOIN_LEFT_OUTER);
-    			PlanNode newChild = createJoinNode();
-    			newChild.setProperty(Info.JOIN_TYPE, JoinType.JOIN_LEFT_OUTER);
-    			joins.remove(childJoin);
-    			
-    			if (childJoin == left) {
-    				newChild.addFirstChild(childJoin.getFirstChild());
-	    			newChild.addLastChild(other);
-	    			newParent.addLastChild(childJoin.getLastChild());
-    			} else {
-    				newChild.addFirstChild(other);
-					newChild.addLastChild(childJoin.getLastChild());
-					newParent.addLastChild(childJoin.getFirstChild());
-    			}
-				newChild.setProperty(Info.JOIN_CRITERIA, joinCriteria);
-				newParent.addFirstChild(newChild);
-				newParent.setProperty(Info.JOIN_CRITERIA, childJoinCriteria);
-				join.getParent().replaceChild(join, newParent);
-    		}
-    	}
-	}
-    
-    private boolean isCriteriaValid(List<Criteria> joinCriteria, QueryMetadataInterface metadata, PlanNode join) {
-    	if (joinCriteria.isEmpty()) {
-    		return false;
-    	}
-		Set<GroupSymbol> groups = join.getGroups();
-    	for (Criteria crit : joinCriteria) {
-			if (JoinUtil.isNullDependent(metadata, groups, crit)) {
-				return false;
-			}
-		}
-    	return true;
-    }
-    
-    private boolean isLeftOuterJoin(PlanNode left) {
-    	return left.getType() == NodeConstants.Types.JOIN && left.getProperty(Info.JOIN_TYPE) == JoinType.JOIN_LEFT_OUTER;
     }
 
     /**
